Add adapter for selection of contests for audit.
