+++ client/src/action/selectContestsForAudit.ts
@@ -4,6 +4,8 @@ import { apiHost } from '../config';
 
 import createSubmitAction from './createSubmitAction';
 
+import adapter from '../adapter/contestsForAudit';
+
 
 const url = `http://${apiHost}/select-contests`;
 
@@ -16,4 +18,5 @@ const selectContestsForAudit = createSubmitAction({
 });
 
 
-export default selectContestsForAudit;
+export default (data: any) =>
+    selectContestsForAudit(adapter(data));
